split challenge_demo.py out of challenge.py
