+++ python/todo.md
@@ -3,4 +3,3 @@ Todo
 
 - wiki template
 - create multiple evaluations?
-- use boto to start scoring machine and install scoring script
